early access features opt in: Fix parsing limitScopeToNetworks response
With limit_scope_to_networks specified,
a subsequent "terraform apply" shows a required change,
as the API returns a list of maps like
`[{"id": "L_1234", "name": "net name"}]`
instead of the list of strings like `["L_1234"]`
which it takes in the request.

Change `fromBody()` manually to convert the list of maps
back to the list of strings -
there does not seem to be an existing option in the definition
for the generator to do that automatically.