network_group_policy: traffic_shaping_rules: Fix application value format
When `traffic_shaping_rules[].definitions[i].type`
is `application` or `applicationCategory`,
the API expects `traffic_shaping_rules[].definitions[i].value` to be a map
(e.g. `{"id": "meraki:layer7/application/171"}`) and fails if it isn't:

        Failed to configure object (PUT), got error: HTTP Request failed: StatusCode
        400, JSON error: ["'value' must be an object"], {"errors":["'value' must be
        an object"]}

That is documented in the property's description in OpenAPI,
which is also copied to the provider's attribute description.
The provider attribute type is `String`, though.

Keep accepting it as a string, but make it a map in API requests
when `type` is `application` or `applicationCategory`,
as was already done for `appliance_l7_firewall_rules`
and `wireless_ssid_traffic_shaping_rules`.