wireless_ssid_l7_firewall_rules: Fix application value format
`rules[i].type` is `application` or `applicationCategory`,
the API expects `rules[i].value` to be a map
(e.g. `{"id": "meraki:layer7/application/171"}`) and fails if it isn't:

        Failed to configure object (PUT), got error: HTTP Request failed: StatusCode
        400, JSON error: ["At least one of your firewall rules is invalid:
        \"network[firewall_rules][3][value] 'value' must be a hash\",
        \"network[firewall_rules][4][value] 'value' must be a hash\"."],
        {"errors":["At least one of your firewall rules is invalid:
        \"network[firewall_rules][3][value] 'value' must be a hash\",
        \"network[firewall_rules][4][value] 'value' must be a hash\"."]}

That is only implicitly documented
in the property's description in OpenAPI
("Format of the value varies
depending on type of the firewall rule selected.")
The provider attribute type is `String`, though.

Keep accepting it as a string, but make it a map in API requests
when `type` is `application` or `applicationCategory`,
as was already done for `appliance_l7_firewall_rules`
and `wireless_ssid_traffic_shaping_rules`.