Fix deletion for single-array-only resources
Singleton resources (ones without a DELETE endpoint)
don't do anything on delete,
leaving the infrastructure still configured.
Aside from being unexpected,
this also sometimes blocks deletion of other resources
(like in https://github.com/CiscoDevNet/terraform-provider-meraki/issues/82).

Some resources (mostly ones with a single "rules" list attribute)
were changed to instead do a PUT with an empty list,
fixing the issue for those resources.

Do that for the rest of such resources
(ones with a single attribute that is a list of maps).

Note that 79 resources that do nothing on delete
still remain after this change.
The issue will have to be documented
and the resources fixed manually on a case-by-case basis
by inferring the expected behavior from the UI.