switch_port: Unset access policy, adaptive policy, profile on deletion
If `access_policy_number` refers to an access policy,
deletion of the access policy is blocked
even after deleting the switch port:
```
│ Error: Client Error
│
│ Failed to delete object (DELETE), got error: HTTP Request failed: StatusCode 400, {"errors":["This access policy is in use by 3 ports. An access policy can only
│ be removed if it is not in use."]}
```

If `adaptive_policy_group_id` refers to an adaptive policy group,
deletion of the group is blocked similarly:
```
│ Error: Client Error
│
│ Failed to delete object (DELETE), got error: HTTP Request failed: StatusCode 400, {"errors":["This adaptive policy group is in use by 8 ports on
│ netascode-network-01 - switch. Adaptive policy groups cannot be removed when in use."]}
```

This is due to meraki_switch_port doing nothing on deletion
(there is no DELETE endpoint).

Do a PUT on deletion, setting these properties:
- `"accessPolicyType": "Open"` - to remove `accessPolicyNumber`.
  Setting `accessPolicyNumber` itself to `null` fails with
  "'accessPolicyNumber' must be an integer".
  Setting the type to `"Open"` does remove the number as well, though.
- `"adaptivePolicyGroupId": null`.
- `"profile": {"enabled": false}`.
  Note: the profile behavior was not tested.
  The `profile_* ` attributes might not be
  blocking the profile's deletion at all
  (and therefore the change might be unnecessary),
  or setting `enabled` to `false` might not be enough
  to avoid `profile_id` possibly blocking the profile's deletion.

`port_schedule_id` does not block deletion of the port schedule -
deleting the port schedule clears the ID in the port;
so it was left unchanged (not cleared on port deletion).