switch_port: Unset adaptive_policy_group_id only if set
Always sending `"adaptivePolicyGroupId": null` on deletion
can fail with "Adaptive Policy is not enabled in this network".
That refers to meraki_organization_adaptive_policy_settings resource
having adaptive policy enabled for this network.

Only send that if `adaptive_policy_group_id` was previously set
(which would mean that adaptive policy
would have been enabled on the network -
we can't check that directly here).