Update UserAgent to use dynamic versioning
This commit modifies the UserAgent string in the provider code to dynamically reflect the version of the MerakiTerraform provider. The hardcoded version has been replaced with a formatted string that utilizes the provider's version attribute. Additionally, the obsolete update-useragent.go script has been removed as its functionality is now integrated into the GitHub Actions workflow.