appliance_vlan_dhcp: Fix definition to correct getPath
rest_endpoint is the same as appliance_vlan's
(/networks/%v/appliance/vlans),
but this resource needs to do a PUT to an existing vlan.

This was done with `put_create: true`,
but the generated Create, Read, Update methods
were using the plural URL (vlans instead of vlans/%v).
This was worked around by manually modifying the methods
to add the ID to the URL.

Now that the methods don't need to be customized for other reasons,
change rest_endpoint and the ID field's definition
to have the generated getPath() handle that instead
in the generated Create, Read, Update methods.