appliance_vlan: Remove PUT-only attributes
Creating a `meraki_appliance_vlan` with `fixed_ip_assignments` set
does not result in `fixedIpAssignments` being set in the API.
A subsequent `terraform apply`
(which does show the pending change again) does result in it being set.

The API ignores the attribute when passed in in the POST request
(which is done on Create) and applies it when passed in the PUT request
(which is done on Update in the subsequent `terraform apply`).
The OpenAPI spec includes the attribute only in the PUT request schema
(though ignoring it instead of returning an error is still unexpected).

`meraki_appliance_vlan_dhcp` does a PUT request
both on Create and Update, so using that resource instead
has consistent behavior.

Remove all the attributes that only exist in the PUT request
to prevent confusion when they are not applied on Create.