appliance_vlan_dhcp_test: Work around failing to create without config
With the customized definition YAML + resource code,
`vlan_id` was not considered an ID field,
so creating a `meraki_appliance_vlan_dhcp`
with just `network_id` and `vlan_id` succeeded.

With the more regular definition YAML and purely generated code,
`vlan_id` is considered an ID field,
so doing the above fails with

        Failed to configure object (POST/PUT), got error: HTTP Request failed:
        StatusCode 400, JSON error: ["Error for
        /networks/{networkId}/appliance/vlans/{vlanId}: None of the fields ('name',
        'subnet', 'applianceIp', 'groupPolicyId', 'vpnNatSubnet', 'dhcpHandling',
        'dhcpRelayServerIps', 'dhcpLeaseTime', 'dhcpBootOptionsEnabled',
        'dhcpBootNextServer', 'dhcpBootFilename', 'fixedIpAssignments',
        'reservedIpRanges', 'dnsNameservers', 'dhcpOptions', 'templateVlanType',
        'cidr', 'mask', 'ipv6' or 'mandatoryDhcp') were specified."],
        {"errors":["Error for /networks/{networkId}/appliance/vlans/{vlanId}: None of
        the fields ('name', 'subnet', 'applianceIp', 'groupPolicyId', 'vpnNatSubnet',
        'dhcpHandling', 'dhcpRelayServerIps', 'dhcpLeaseTime',
        'dhcpBootOptionsEnabled', 'dhcpBootNextServer', 'dhcpBootFilename',
        'fixedIpAssignments', 'reservedIpRanges', 'dnsNameservers', 'dhcpOptions',
        'templateVlanType', 'cidr', 'mask', 'ipv6' or 'mandatoryDhcp') were
        specified."]}

which presumably happens for other singletons as well.

Other resources' tests deal with this either
by having at least one field with `minimum_test_value`
in the definition YAML
or by having `skip_minimum_test: true` there.

Change the test so that one config field is passed in the minimal test.