Generated toBody: Allow specifying empty lists of maps
Currently, toBody() does not distinuish between list-of-maps attributes
being null and empty.
This results in the provider omitting the attribute from the API request
when the value is an empty list,
making it impossible for the user to clear a list of maps.

Handling of lists of primitive values like strings
does not have this issue -
attributes with an empty list value are included in the request
while ones with a null value are skipped.

Change the check to only omit list-of-maps attributes with null values.